refactor: router에 따른 폴더구조 변경
