+++ package.json
@@ -14,7 +14,8 @@
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-id-generator": "^2.0.0",
-    "react-prop-types": "^0.4.0"
+    "react-prop-types": "^0.4.0",
+    "react-router-dom": "^5.0.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.5.5",
