+++ src/About.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function About() {
+  return (
+    <div>
+      어바웃
+    </div>
+  )
+}
